# app/API/HealthTrackerAPI/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HealthTrackerAPI",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5181"
    }
  ],
  "paths": {
    "/api/Activity/stats": {
      "get": {
        "tags": [
          "Activity"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/monthly": {
      "get": {
        "tags": [
          "Activity"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/yoy/{month}": {
      "get": {
        "tags": [
          "Activity"
        ],
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/annual": {
      "get": {
        "tags": [
          "Activity"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/totals": {
      "get": {
        "tags": [
          "Activity"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/date/{date}": {
      "get": {
        "tags": [
          "Activity"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity": {
      "post": {
        "tags": [
          "Activity"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Activity/date/{existingUuid}": {
      "put": {
        "tags": [
          "Activity"
        ],
        "parameters": [
          {
            "name": "existingUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddStepsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/dashboard/latest": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Ping endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/clearapi/push": {
      "get": {
        "tags": [
          "HealthTrackerAPI"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Pedometer": {
      "get": {
        "tags": [
          "Pedometer"
        ],
        "summary": "Returns a paginated, filterable, sortable page of pedometer summaries",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "search_dayTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search_stepCount",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search_calorie",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "dayTime"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Pedometer/stats": {
      "get": {
        "tags": [
          "Pedometer"
        ],
        "summary": "Returns pre-aggregated pedometer stats",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Strings": {
      "get": {
        "tags": [
          "Strings"
        ],
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Strings/audit": {
      "get": {
        "tags": [
          "Strings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Strings/{page}/{uniqueId}": {
      "put": {
        "tags": [
          "Strings"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uniqueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "en"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStringRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStringRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStringRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight": {
      "get": {
        "tags": [
          "Weight"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "search_createTime",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search_weight",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createTime"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Weight"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/stats": {
      "get": {
        "tags": [
          "Weight"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/yoy/{month}": {
      "get": {
        "tags": [
          "Weight"
        ],
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/annual": {
      "get": {
        "tags": [
          "Weight"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/monthly": {
      "get": {
        "tags": [
          "Weight"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/date/{date}": {
      "get": {
        "tags": [
          "Weight"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Weight/date/{existingUuid}": {
      "put": {
        "tags": [
          "Weight"
        ],
        "parameters": [
          {
            "name": "existingUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddStepsRequest": {
        "type": "object",
        "properties": {
          "stepCount": {
            "type": "integer",
            "format": "int32"
          },
          "dayTime": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddWeightRequest": {
        "type": "object",
        "properties": {
          "weightKg": {
            "type": "number",
            "format": "double"
          },
          "createTime": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpsertProfileRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "heightCm": {
            "type": "number",
            "format": "double"
          },
          "heightUom": {
            "type": "string",
            "nullable": true
          },
          "weightUom": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpsertStringRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Activity"
    },
    {
      "name": "Dashboard"
    },
    {
      "name": "Health"
    },
    {
      "name": "HealthTrackerAPI"
    },
    {
      "name": "Pedometer"
    },
    {
      "name": "Profile"
    },
    {
      "name": "Strings"
    },
    {
      "name": "Weight"
    }
  ]
}